The change here is getting committed for DERBY-4048
DERBY-269 provided a new stored procedure SYSCS_UTIL.SYSCS_UPDATE_STATISTICS to update the statistics. The code to do the work was added in AlterTableConstantAction and the name of new method in that class is updateStatistics(Activation activation). This code was copied from an existing unused class execute/UpdateStatisticsConstantAction.java(that unused class has been removed from Derby code). 

The copied code in AlterTableConstantAction:updateStatistics gets compile transaction to do the work of updating the statistics. Instead, we should be using execute transaction because we are not in the compile phase. The rest of the code in AlterTableConstantAction uses execute transaction.

I ran into this issue of getTransactionExecute vs getTransactionCompile while working on DERBY-3788 (provide a zero-admin way of updating the statistics). While trying to run the statistics inline during the compile phase of a SELECT query, I ran into exception that statistics can't be updated because we are in read-only mode. But if I change the AlterTableConstantAction to use getTransactionExecute, then I do not run into the read-only mode exception.  



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@741118 13f79535-47bb-0310-9956-ffa450edef68