DERBY-4059 If space reclamation cannot obtain container lock it will not retry getting the lock
This change does make the change so that the retry occurs but in practical application it is unlikely that we will get the lock after three retries, because the retry happens so quickly. See https://issues.apache.org/jira/browse/DERBY-4055?focusedCommentId=12673436#action_12673436 for suggestions on long term solutions to the retry problem.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@745982 13f79535-47bb-0310-9956-ffa450edef68