DERBY-4009 (partial) Accommodate length delimited DRDA strings where character length does not equal byte length
Change client RDBNAM, USERID, and PASSWORD to allow for CcsidManager where the byte length might not equal the string length. Right now we just have EbcdicCcsidManager which has a maximum bytes per char as 1. But, this lays the groundwork for UTF8 which will be added later as part of DERBY-728 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@747881 13f79535-47bb-0310-9956-ffa450edef68