DERBY-4088: DDMReader readBytes ArrayIndexOutOfBoundsException
When ensureBLayerDataInBuffer() calls compressBLayerData() the number
of bytes in the buffer is reduced (header bytes are stripped out) and
there might be too little data in the buffer when it returns. This
patch fixes the problem by ensuring that compressBLayerData() fills
the buffer with enough extra bytes to compensate for the removal of
the header bytes.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@752813 13f79535-47bb-0310-9956-ffa450edef68