DERBY-4097: BatchUpdateException in NullableUniqueConstraintTest
Improve error reporting when a BatchUpdateException is thrown by the
client driver. Use initCause() to put the underlying exception in the
chain of exceptions printed by printStackTrace(). Note that there may
be more than one underlying exception for a BatchUpdateException and
only one of them will be printed by printStackTrace(). To see the rest
of the exceptions, getNextException() has to be used (as before).


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@755147 13f79535-47bb-0310-9956-ffa450edef68