DERBY-2487: Enhance Derby with XPLAIN functionality.
This feature was contributed by Felix Beyer (beyer dot felix at gmx dot net).

This feature introduces an alternate handling of runtime statistics
information. Derby can now be configured so that it will capture the
statistics in a machine-readable form and will store them into a set of
simply-structured tables in a schema which is specified by the user. We
call this behavior "XPLAIN style", and we call the tables which are
used the "XPLAIN tables".

Having captured statistics about statement execution, you can then
analyze the statement behavior by querying these tables. For example, you
can determine how much time was taken, what resources were used, what
query plan was chosen, and so on.

This is the initial commit of this new functionality. Subsequent changes
will provide documentation, additional tests, fully-tested upgrade support.
As the community gains more experience with this feature, we will find more
enhancements to make, but the initial commit provides a working feature
with an initial set of basic regression tests.

Many thanks to the community for reviewing the changes through several
iterations of development.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@768597 13f79535-47bb-0310-9956-ffa450edef68