DERBY-4186 After master stop, test fails when it succeeds in connecting (rebooting) shut-down ex-slave
Patch derby-4186-2, which solves this issue by

a) fixing a bug in SlaveDataBase, which loses and exception iff a
command to stop replication arrives from the master to the slave
before the initial connection (successful slave started) command
returns. This is a corner case race condition, and not very likely to
occur in practice, since it makes little sense to stop replication
immediately after starting it.

b) fixing a bug in the test, which switched on its head what is really
expected behavior, and also adds a waiting loop to allow intermediate
state on slave after the master is stopped. Also fixed some erroneous
comment and removed some cruft.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@769962 13f79535-47bb-0310-9956-ffa450edef68