DERBY-4193: ASSERT FAILED Scan position already saved with multi-threaded insert/update/delete
Forget about the current position before trying to reposition on the
end point of a scan. Then the ASSERT won't be confused if it needs to
save the position again (it fails if it finds that a position is
already saved).

BTreeScan.positionAtStartForBackwardScan() also had this problem, but
since it doesn't have any callers, this method was removed instead of
being fixed.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@772090 13f79535-47bb-0310-9956-ffa450edef68