DERBY-4062: Remove single-argument getDataValue overrides from DataValueFactory
This change simplifies the DataValueFactory interface by removing the
single-argument overrides of the getDataValue() method. This change extends
the work that was done by Dan in 2007 in revision 540791. The large number
of single-argument overrides was complicated and clients could accidentally
call the wrong override, particularly since there was an override which
took an argument of type Object. As suggested by Knut, the new implementation
replaces calls to the single-argument getDataValue method with direct calls
to the constructor of the desired data type. The new code is simpler and
easier to read, and the data type usage is more self-evident in the system
catalog row factories, which are the principal users of the single-argument
getDataValue calls.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@774617 13f79535-47bb-0310-9956-ffa450edef68