DERBY-4204: Statistics not collected on re-execution of statement
This change enhances the DERBY-4204 regression test to verify that the
XPLAIN table form of statistics collection is also re-collected on each
statement re-execution. Without the DERBY-4204 fix, re-execution of DML
statements was only resulting in a single row in the SYSXPLAIN_STATEMENTS
table, while after the fix, the correct results of 5 rows are observed.

This change should NOT be backported to 10.5 and earlier branches, because
it depends on the XPLAIN functionality.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@774830 13f79535-47bb-0310-9956-ffa450edef68