DERBY-4182
Before this fix abort of inserts that included clob or blob chains would 
destroy the links of the allocated pages of the chains.  This would leave
allocated pages that could never be reclaimed either by subsequent post
commit processing or inplace compress.  Only offline compress could reclaim
the space.  This fix changes insert abort processing to automatically put
all pieces of long columns except for the head page on the free list as part
of the abort.

Note this does not fix existing tables that have had this problem happen in
the past, only stops it from happening.  One must run an offline compress to
reclaim this space to fix any instances of this bug prior to this fix.




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@778926 13f79535-47bb-0310-9956-ffa450edef68