DERBY-4239
backporting svn #780352 from trunk to 10.1 branch.

In the case of inplace compress it is necessary to get a valid checkpoint
into the log file after any other log record for the table being compressed.
This insures that if redo recovery is run, it will not be applied to any
log record before the compress. This is necessary as compress will basically
destroy a page and redo recovery will get various errors if it tries to
apply a redo page operation at page version N against a new page.

Before this fix the requested checkpoint would not be executed if a checkpoint
was already in progress. The problem is that there is no way to tell if
the executing checkpoint is after all operations on the table. So the fix
in this case waits for the checkpoint to finish and starts a new one.

The changes also apply this new behavior to all calls to checkpoint which
previously had marked that the routine should wait until the checkpoint
finished. This will affect backup and checkpoints initiated by a user
calling the system procedure. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.1@782042 13f79535-47bb-0310-9956-ffa450edef68