DERBY-4047
updating the master and adding the test to be run as part of the storemats
suite.  I have checked the diffs with the neglected master and found them
all to be ok.

I hand checked the diffs and found the following 3 family of diffs to be
acceptable:
o a bunch with the new ij header
o trunk and 10.5 does not have scan locks, so no longer will get any row locks
of the form (N, 1)
o the following kind of diff comes from inlist optimization, now a set of
OR's can do multiple index probes and thus only the affected rows are
locked.  The exact number and name of the rows depends on the type of
index and the padding of the row.
> -- Very basic single user testing of update locks.  This ".subsql" test is
3417,3418d3521
< APP     |UserTran|ROW     |1   |X   |A           |(1,11)    |GRANT|ACTIVE
< APP     |UserTran|ROW     |1   |X   |A           |(1,9)     |GRANT|ACTIVE
3634,3638d3736
< APP     |UserTran|ROW     |1   |X   |A           |(1,10)    |GRANT|ACTIVE
< APP     |UserTran|ROW     |1   |X   |A           |(1,11)    |GRANT|ACTIVE
< APP     |UserTran|ROW     |1   |X   |A           |(1,12)    |GRANT|ACTIVE
< APP     |UserTran|ROW     |1   |X   |A           |(1,8)     |GRANT|ACTIVE
< APP     |UserTran|ROW     |1   |X   |A           |(1,9)     |GRANT|ACTIVE

delete from a where a = 2 or a = 4 or a = 6;
2 rows inserted/updated/deleted
ij> select * from lock_table order by tabname, type desc, mode, cnt, lockname;
USERNAME|TRANTYPE|TYPE    |CNT |MODE|TABNAME     |LOCKNAME  |STATE|STATUS
---------------------------------------------------------------------------
APP     |UserTran|TABLE   |2   |IX  |A           |Tablelock |GRANT|ACTIVE
APP     |UserTran|ROW     |1   |X   |A           |(1,10)    |GRANT|ACTIVE
APP     |UserTran|ROW     |2   |X   |A           |(1,12)    |GRANT|ACTIVE
APP     |UserTran|ROW     |2   |X   |A           |(1,8)     |GRANT|ACTIVE



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@783088 13f79535-47bb-0310-9956-ffa450edef68