DERBY-4198 When using the FOR UPDATE OF clause with SUR (Scroll-insensive updatable result sets), the updateRow() method crashes
Patch derby-4198-4. This is the second part of the solution for this
issue. It solves the issue of column mapping in the presence of named
columns in a FOR UPDATE OF clause. The original code was not general
enough; ScrollInsensitiveResultSet.updateRow needs to make use of
RowChanger to do the right thing.

The patch also adds new test cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@785163 13f79535-47bb-0310-9956-ffa450edef68