DERBY-4304
When running in insane mode, the stack trace of the exception was not getting printed in the console log file logs/serverConsoleOutput.log The reason for this is that the code was
doing following
            consolePropertyMessage("DRDA_UnexpectedException.S",			
                exception.getMessage());
            consoleExceptionPrintTrace(exception);
That is, the stack trace printing was happening after the call to consolePropertyMessage. The call to consolePropertyMessage results into a call to consolePropertyMessageWork which may throw an exception back depending on the type of the error it is handling. If this method does throw an exception, then consoleExceptionPrintTrace does not get a chance to dump the stack trace into the logs/serverConsoleOutput. (Probably in sane mode, there is some other additional place where we do the stack trace printing and that is why we saw the stack trace despite consolePropertyMessage throwing an exception.) To fix the problem, I have switched the order of the 2 calls above and that makes sure that we do print the stack trace for debugging purpose. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@796316 13f79535-47bb-0310-9956-ffa450edef68