DERBY-4248: Convert checkConstraint.sql to JUnit
This patch was contributed by Eranda Sooriyabandara (070468D at gmail dot com)

This patch contains a re-written copy of the checkConstraint test suite.
checkConstraint.sql was converted to JUnit format using the SQLToJUnit tool,
and the resulting Java program was finished by hand by Eranda. The patch
removes the old checkConstraint.sql and checkConstraint.out files, and
also removes checkConstraint.sql from the old derbylang suite. The new
CheckConstraintTest is added to the JUnit lang._Suite.

There is currently a portion of this test which is commented out, because
it provokes DERBY-4282. When DERBY-4282 is resolved, that portion of the
test should be re-enabled.


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@796893 13f79535-47bb-0310-9956-ffa450edef68