DERBY-4306
registeredMbeans is null but we do not check for nullability before executing following
if (registeredMbeans.remove(mbeanName) == null) 
which resulted in NPE. 

I have fixed the problem by checking for nullability first. registeredMbeans will be null if someone shuts down the engine first and then later on issues a command to shutdown the server. During the shutdown, registeredMbeans will be null and we should check for that nullability before accessing it.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@797147 13f79535-47bb-0310-9956-ffa450edef68