Merging 797147 from trunk into 10.5 codeline. We make a check for nullability before accessing registeredMbeans. The commit comments for revision 797147 are as follows
******************** 
DERBY-4306 

registeredMbeans is null but we do not check for nullability before executing following 
if (registeredMbeans.remove(mbeanName) == null) 
which resulted in NPE. 

I have fixed the problem by checking for nullability first. registeredMbeans will be null if someone shuts down the engine first and then later on issues a command to shutdown the server. During the shutdown, registeredMbeans will be null and we should check for that nullability before accessing it. 
******************** 


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.5@797434 13f79535-47bb-0310-9956-ffa450edef68