DERBY-4304
During one of the server shutdown steps, DERBY-4306 was throwing a npe(we have fixed DERBY-4306 now) and the shutdown exception handling code was calling following method on that npe 
          consolePropertyMessage("DRDA_UnexpectedException.S", exception.getMessage()); 
          consoleExceptionPrintTrace(exception); 
but the call exception.getMessage() on npe was returning null and consolePropertyMessage is not written to handle null 2nd param for this call of consolePropertyMessage. But the work done by this particular type of consolePropertyMessage method is also done by the next call in the shutdown exception handling which is consoleExceptionPrintTrace. Because of this, I have replaced the consolePropertyMessage("DRDA_UnexpectedException.S", exception.getMessage()); call with consolePropertyMessage("DRDA_UnexpectedException.S", true); This change makes sure we do not run into npe when exception.getMessage() is null. I am not sure if we should enter a new jira entry to make consolePropertyMessage("DRDA_UnexpectedException.S", exception.getMessage()); handle a null 2nd param. 

junit tests with this change ran into intermittent upgrade test failures. 

One last step remaining is may be write a new method which is a subset of consolePropertyMessage because consolePropertyMessage handles the exception printing and then goes ahead and throws the exception again. For our purposes in this jira entry, we do not want the exception to be thrown back. instead, we want to move on to the next step in server shutdown. 



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@798347 13f79535-47bb-0310-9956-ffa450edef68