DERBY-4256: Allow ALTER TABLE to increase the max size of BLOB or CLOB
This patch was contributed by Eranda Sooriyabandara (070468D at gmail dot com)

This patch enhances the Derby ALTER TABLE command so that it allows the
maximum length limit of a BLOB or CLOB column to be increased. The patch also
adds some new regression tests verifying the new functionality, and adjusts
the text of the ALTER TABLE error message to make it more clear that it is
legal to increase the length of a VARCHAR, BLOB, or CLOB column.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@802985 13f79535-47bb-0310-9956-ffa450edef68