DERBY-4304
Committing changes which puts a new try catch block around all of the server shutdown code in case if there is some exception being thrown which is not getting caught. The catch block will log into the log file but if that fails for some reason, it will also just dump the stack trace using ex.printStackTrace(). 




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@803548 13f79535-47bb-0310-9956-ffa450edef68