DERBY-4330 NullPointerException or assert failure when re-executing PreparedStatement after lock timeout
Backported fix from trunk as:
svn merge -c 804271 https://svn.apache.org/repos/asf/db/derby/code/trunk

Patch derby-4330c fixes this issue. The problem is that when a timeout
happens (or a deadlock), the result set tree for prepared statements
for some queries is partically in a closed, partially in an open
state. (The issue was reported for a join query, but exists for others
queries as well). This causes problems when the result set tree is
being reused, i.e. when the prpared statement is attempted re-executed
after the timeout, since the tree is expected to be fully closed at
that time, cuasing the assert or NPE.

The fix ensures that the tree is left in a fully closed state in such
cases.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/branches/10.4@804327 13f79535-47bb-0310-9956-ffa450edef68