DERBY-4081: BTreeController.comparePreviousRecord() may fail to release latch on left-most leaf
When a new value is inserted into a nullable unique index, the rows
immediately to the left and the right of the insertion point are
checked for duplicates. comparePreviousRecord() checks the rows to the
left of the insertion point. If the row immediately to the left is
marked as deleted, it will have to go further to the left until it
finds a row that is not deleted.

Now this check may cross page boundaries, and if it does so, and all
the rows to the left of the insertion point are deleted,
comparePreviousRecord() will return NO_MATCH before the latch
left-most leaf in the index has been released.

This means that the left-most leaf is still latched when
comparePreviousRecord() returns, which it is not supposed to be unless
the new value was actually inserted into the left-most leaf. The
unexpected extra latch may cause deadlocks or livelocks.

This patch adds a test that runs into a livelock because of the bug,
and it fixes the bug by releasing the latch on the left-most leaf
before returning NO_MATCH from comparePreviousRecord().


git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@805696 13f79535-47bb-0310-9956-ffa450edef68