DERBY-4087 Clean up debug printing of the abstract syntax trees after parsing, binding and optimization
Patch DERBY-4087-b, which improves debug-time printing of query trees.
Much of the changes concerns cleaning up when to use toString and when
to use printSubNodes, a pattern which is well-defined but was severely
broken, as well as including more data for each node, and sub-trees
not hitherto printed at all. Please see JIRA for mroe detail.

A new feature highlights aliases by suppressing printing of already
printed subtrees (the query tree is a DAG, not a pure tree).




git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@808523 13f79535-47bb-0310-9956-ffa450edef68