DERBY-4284: All Columns become Nullable when Using left join
Only the columns from the right side of a left outer join, or
from the left side of a right outer join, should be changed to
nullable. This is done correctly in JoinNode.buildRCL(), but in
addition SelectNode.bindResultColumns() makes all result columns
nullable if the from list contains an outer join.

This patch fixes the bug by removing the code in SelectNode that
makes all columns nullable. This also made it possible to remove
a method in FromList that became unused after the fix. Also, some
test canons needed to be updated because IJ may change the
formatting/spacing slightly depending on whether the result
columns are nullable or not.

The fix also exposed another manifestation of DERBY-2916, where
some nodes saw the wrong nullability because they looked at the
columns before JoinNode.buildRCL() had done its job. This was
addressed by adding a workaround to JoinNode.getMatchingColumn().

Also, changes that had been made to wisconsin.out as part of
DERBY-2916 earlier in order to make the Wisconsin test run
cleanly despite the bug, could now be reverted because of the
workaround.

git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@810860 13f79535-47bb-0310-9956-ffa450edef68