DERBY-4361
It was not enough to just the shutdown the database in Windows in order to be able to delete
the log file later. The engine needed to be brought down too. I added code in JDBCDataSource for shutting down the engine and used this new method from the ClassLoaderBootTest to shut the engine down.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@812669 13f79535-47bb-0310-9956-ffa450edef68