DERBY-4363: Add simple benchmark for measuring GROUP BY performance.
This change introduces a simple benchmark which can be used to measure
GROUP BY statement performance. As an example of how to run the benchmark:

   java org.apache.derbyTesting.perf.clients.Runner -load group_by -init
   -load_opts numGroupingCols=2,numGroups1=10,numGroups2=100
   
(all on one long line). 

Over time, I anticipate that we will add other capabilities to this test to
measure other aspects of GROUP BY performance, but the benchmark seems
sufficiently functional at this point to be worth checking in.



git-svn-id: https://svn.apache.org/repos/asf/db/derby/code/trunk@814237 13f79535-47bb-0310-9956-ffa450edef68